@@ -17,9 +17,10 @@
     "lint:fix": "eslint */**/*.{ts,tsx} --fix",
     "lint:quiet": "eslint */**/*.{ts,tsx} --quiet"
   },
+  "license": "MIT",
   "dependencies": {
     "@azure/communication-chat": "^1.4.0",
-    "@azure/communication-common": "^2.3.0",
+    "@azure/communication-common": "^2.3.1",
     "@azure/communication-identity": "^1.3.0",
     "@azure/storage-blob": "^12.15.0",
     "@types/cors": "^2.8.8",
@@ -38,33 +39,33 @@
     "@types/copy-webpack-plugin": "^6.4.0",
     "@types/express": "^4.17.8",
     "@types/http-errors": "^1.8.0",
-    "@types/jest": "~26.0.22",
+    "@types/jest": "^29.5.11",
     "@types/morgan": "^1.9.2",
-    "@types/node": "^14.14.10",
-    "@types/supertest": "^2.0.10",
-    "@types/webpack-node-externals": "^2.5.0",
-    "@typescript-eslint/eslint-plugin": "^4.12.0",
-    "@typescript-eslint/parser": "^4.12.0",
-    "copy-webpack-plugin": "^6.4.0",
+    "@types/node": "^20.11.25",
+    "@types/supertest": "^2.0.12",
+    "@types/webpack-node-externals": "^3.0.4",
+    "@typescript-eslint/eslint-plugin": "^6.19.1",
+    "@typescript-eslint/parser": "^6.19.1",
+    "copy-webpack-plugin": "^11.0.0",
     "eslint": "^7.7.0",
     "eslint-config-prettier": "^6.12.0",
     "eslint-plugin-flowtype": "^5.2.0",
-    "eslint-plugin-import": "~2.22.1",
+    "eslint-plugin-import": "^2.29.0",
     "eslint-plugin-prettier": "^3.1.4",
-    "husky": "^4.3.0",
-    "jest": "26.6.0",
-    "jest-junit": "~13.0.0",
+    "husky": "^8.0.3",
+    "jest": "29.7.0",
+    "jest-junit": "^16.0.0",
     "prettier": "2.3.1",
     "pretty-quick": "^3.1.0",
     "rimraf": "^2.6.2",
-    "supertest": "^6.0.1",
-    "ts-jest": "^26.4.4",
+    "supertest": "^6.3.3",
+    "ts-jest": "^29.1.2",
     "ts-loader": "^8.0.12",
-    "ts-node": "^9.1.1",
-    "ts-node-dev": "^1.0.0",
-    "typescript": "4.3.5",
-    "webpack": "5.61.0",
-    "webpack-cli": "4.7.2",
-    "webpack-node-externals": "^2.5.2"
+    "ts-node": "^10.9.2",
+    "ts-node-dev": "^2.0.0",
+    "typescript": "5.4.2",
+    "webpack": "5.89.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-node-externals": "^3.0.0"
   }
 }
