@@ -81,7 +81,7 @@ export const ChatScreen = (props: ChatScreenProps): JSX.Element => {
   }, [adapter]);
 
   if (adapter) {
-    const onFetchAvatarPersonaData = (userId): Promise<AvatarPersonaData> =>
+    const onFetchAvatarPersonaData = (userId: string): Promise<AvatarPersonaData> =>
       fetchEmojiForUser(userId).then(
         (emoji) =>
           new Promise((resolve) => {
